requirements.txt
aniso8601                         9.0.1
backports.entry-points-selectable 1.1.1
blinker                           1.4
cachelib                          0.6.0
certifi                           2021.10.8
cffi                              1.15.0
charset-normalizer                2.0.10
click                             8.0.3
colorama                          0.4.4
cryptography                      36.0.1
Deprecated                        1.2.13
distlib                           0.3.4
filelock                          3.4.0
Flask                             2.0.2
Flask-Cors                        3.0.10
Flask-Mail                        0.9.1
Flask-RESTful                     0.3.8
Flask-Session                     0.4.0
Flask-SQLAlchemy                  2.5.1
greenlet                          1.1.2
idna                              3.3
itsdangerous                      2.0.1
Jinja2                            3.0.3
MarkupSafe                        2.0.1
numpy                             1.22.2
pandas                            1.4.0
pip                               21.2.4
platformdirs                      2.4.0
psycopg2                          2.9.3
pycparser                         2.21
PyJWT                             1.7.1
pyotp                             2.6.0
python-dateutil                   2.8.2
pytz                              2021.3
PyYAML                            5.4.1
requests                          2.27.1
setuptools                        58.1.0
six                               1.16.0
SQLAlchemy                        1.4.29
urllib3                           1.26.8
virtualenv                        20.10.0
vonage                            2.5.5
Werkzeug                          2.0.2
wrapt                             1.13.3